+++ CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
  * Remove `nikic/php-parser` dependency
  * Add `_current_object` to the context passed to value transformers during write operations
  * Add `include_null_properties` option to encode the properties with `null` value
+ * Add synthetic properties support
 
 7.3
 ---